Updated: sphere.ini SectorSleep description. Fixed: Exception when trying to retrieve specific map blocks. Fixed: Map and UID caches reserved too much memory. Small performance improvements in CPoint* methods. Changed CMultiStorage underlying containers. Fixed Linux compilation error. Fixed a couple more of uninitialized members.

+++ src/game/items/CItem.cpp
@@ -3443,7 +3443,7 @@ void CItem::SetTriggerActive(lpctstr trig)
     int iAction = FindTableSorted( trig, CItem::sm_szTrigName, CountOf(CItem::sm_szTrigName)-1 );
     if (iAction != -1)
     {
-        _iRunningTriggerId = iAction;
+        _iRunningTriggerId = (short)iAction;
         _sRunningTrigger.Empty();
         return;
     }
